major update from dev branch:
- fix order of plotting for alpha value when showing all test curves
- require exact parameter matches (fix #12)
- more general naming (fix #15)
- more general legend info for non-optimized params (fix #13)
- simple option to incorporate old results (fix #14)
- minor change to simplify example input file for easier testing

+++ opt_fea.py
@@ -56,11 +56,14 @@ def loop(opt, loop_len):
                     return
                 else:
                     job_extract()  # extract data to temp_time_disp_force.csv
-                    combine_SS(zeros=False)  # save stress-strain data
-                    rmse = calc_error()  # get error
-                    opt.tell(next_params, rmse)
-                    opt_progress = update_progress(i, next_params, rmse)
-                    write_opt_progress()
+                    if np.sum(np.loadtxt('temp_time_disp_force.csv', delimiter=',', skiprows=1)[:,1:2]) == 0:
+                        write_maxRMSE(i, next_params, opt)
+                    else:
+                        combine_SS(zeros=False)  # save stress-strain data
+                        rmse = calc_error()  # get error
+                        opt.tell(next_params, rmse)
+                        opt_progress = update_progress(i, next_params, rmse)
+                        write_opt_progress()
         single_loop(opt, i)
 
 
@@ -154,8 +157,10 @@ def write_opt_progress():
 
 def update_progress(i, next_params, rmse):
     global opt_progress
-    if i == 0: opt_progress = np.transpose(np.asarray([i] + next_params + [rmse]))
-    else: opt_progress = np.vstack((opt_progress, np.asarray([i] + next_params + [rmse])))
+    if (i == 0) and (uset.do_load_previous == False): 
+        opt_progress = np.transpose(np.asarray([i] + next_params + [rmse]))
+    else: 
+        opt_progress = np.vstack((opt_progress, np.asarray([i] + next_params + [rmse])))
     return opt_progress
 
 
@@ -198,25 +203,31 @@ def get_first():
 def load_opt(opt):
     """
     Load input files of previous optimizations to use as initial points in current optimization.
-    Note the parameter bounds for `in`-files must be within current bounds.
+    Note the parameter bounds for the input files must be within current parameter bounds.
     """
-    filename = 'in_opt.txt'
-    arrayname = 'in_opt.npy'
-    if os.path.isfile(filename):
-        prev_data = np.loadtxt(filename, skiprows=1)
-        x_in = prev_data[:,1:-1].tolist()
-        y_in = prev_data[:,-1].tolist()
+    global opt_progress
+    filename = 'out_progress.txt'
+    arrayname = 'out_time_disp_force.npy'
+    if uset.do_load_previous:
+        opt_progress = np.loadtxt(filename, skiprows=1)
+        # renumber iterations (negative length to zero) to distinguish from new calculations:
+        opt_progress[:,0] = np.array([i for i in range(-1*len(opt_progress[:,0]),0)])
+        x_in = opt_progress[:,1:-1].tolist()
+        y_in = opt_progress[:,-1].tolist()
         opt.tell(x_in, y_in)
     if os.path.isfile(arrayname):
         np.save(arrayname, np.load(arrayname))
 
 
 def remove_out_files():
-    out_files = [f for f in os.listdir(os.getcwd()) \
-        if (f.startswith('out_') or f.startswith('res_') or f.startswith('temp_'))]
-    if len(out_files) > 0:
-        for f in out_files:
-            os.remove(f)
+    if not uset.do_load_previous:
+        out_files = [f for f in os.listdir(os.getcwd()) \
+            if (f.startswith('out_') or f.startswith('res_') or f.startswith('temp_'))]
+        if len(out_files) > 0:
+            for f in out_files:
+                os.remove(f)
+    job_files = [f for f in os.listdir(os.getcwd()) \
+        if (f.startswith(uset.jobname)) and not (f.endswith('.inp'))]
 
 
 def param_check(param_list):
@@ -250,7 +261,7 @@ def max_rmse(loop_number):
             return uset.large_error
         else:
             iq1, iq3 = np.quantile(errors, [0.25,0.75])
-            return (iq3-iq1)*1.5
+            return np.ceil(np.mean(errors) + (iq3-iq1)*1.5)
 
 
 def check_complete():
@@ -404,7 +415,7 @@ def write_parameters(next_params):
         for line in lines:
             skip = False
             for param in uset.param_list:
-                if line.startswith(param):
+                if line[:line.find('=')].strip() == param:
                     f2.write(param + ' = ' + str(next_params[uset.param_list.index(param)]) + '\n')
                     skip = True
             if not skip:
